google-ai-studio/requirements.txt
google-generativeai
google-genai
typer
rich
